Deleting the old (unused) analysis and expand code
