+++ proc_blocks/macros/src/lib.rs
@@ -5,8 +5,8 @@ extern crate alloc;
 #[macro_use]
 extern crate pretty_assertions;
 
-mod analysis_2;
-mod expand_2;
+mod analysis;
+mod expand;
 mod types;
 
 // This is a bit hacky, but by using a #[path] attribute we can share the
@@ -130,7 +130,7 @@ use syn::DeriveInput;
 pub fn proc_block(input: TokenStream) -> TokenStream {
     let input = syn::parse_macro_input!(input as DeriveInput);
 
-    let tokens = analysis_2::analyse(&input)
+    let tokens = analysis::analyse(&input)
         .map(ToTokens::into_token_stream)
         .unwrap_or_else(|e| e.into_compile_error().into());
